drm/vc4: Replace HDMI force-connected with an EDID probe.
The force-connected started out because I didn't know how to read the
HPD pin successfully, which required the hpd_active_low check and
getting the correct active level into the DTs.  It stayed because we
don't have the Pi3's HPD line exposed to Linux, so this was the only
way to bring up graphics on it.

However, with the DSI panel support now present, users want to be able
to run DSI-only systems, and forcing HDMI on is interfering with
default screen configurations.  Work around the Pi3's missing HPD by
probing the DDC on I2C and see if it's present at all.

Signed-off-by: Eric Anholt <eric@anholt.net>